CAN-74: Manage Profile Frontend Design Fix 2
CAN-74: Manage Profile Frontend Design Fix 2
